Take percentage widths/heights into account
When using percentages as widths or heights, the existing code returns an Integer value through parseInt (100% becomes 100), resulting in an incorrect video aspect ratio. I've added a check for the '%' character in the conditional statements, so that when percentage widths and height are used, the actual height or width of the object on the screen is measured by jQuery (using width() and height() functions).